MSDemandForecastEngine/requirements.txt
functions_framework==3.0.0
gcsfs==2024.6.1
google-cloud-secret-manager==2.19.0
google-cloud-pubsub==2.22.0
google-cloud-bigquery==3.30.0
pandas==2.2.2
pyarrow==16.1.0
fastparquet==2023.2.0
pymssql==2.3.1
db-dtypes==1.4.1
google-cloud-storage==3.1.0
numpy==1.26.4
xgboost==3.0.2
statsmodels==0.14.4
scikit-learn == 1.6.1